+++ Civitas.Id.Sweden.Json/Options/ValidateCivitasIdSwedenJsonOptions.cs
@@ -1,0 +1,7 @@
+using Microsoft.Extensions.Options;
+
+namespace Civitas.Id.Sweden.Json.Options;
+
+[OptionsValidator]
+internal sealed partial class ValidateCivitasIdSwedenJsonOptions
+    : IValidateOptions<CivitasIdSwedenJsonOptions>;